chore: move pages to separate folder
