+++ i18n/en.pot
@@ -5,23 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2019-06-18T14:42:06.050Z\n"
-"PO-Revision-Date: 2019-06-18T14:42:06.050Z\n"
-
-msgid "Are you sure you want to discard your changes?"
-msgstr ""
-
-msgid "Add new job"
-msgstr ""
-
-msgid "Attributes"
-msgstr ""
-
-msgid "Name"
-msgstr ""
-
-msgid "Job type"
-msgstr ""
+"POT-Creation-Date: 2019-06-19T13:44:29.997Z\n"
+"PO-Revision-Date: 2019-06-19T13:44:29.997Z\n"
 
 msgid "Scheduler"
 msgstr ""
@@ -39,6 +24,36 @@ msgid "Confirm"
 msgstr ""
 
 msgid "Open user guide"
+msgstr ""
+
+msgid "Select frequency"
+msgstr ""
+
+msgid "Cron expression"
+msgstr ""
+
+msgid "Continuous execution"
+msgstr ""
+
+msgid "Name"
+msgstr ""
+
+msgid "Job type"
+msgstr ""
+
+msgid "Click to select"
+msgstr ""
+
+msgid "Parameters"
+msgstr ""
+
+msgid "Are you sure you want to discard your changes?"
+msgstr ""
+
+msgid "Add new job"
+msgstr ""
+
+msgid "Attributes"
 msgstr ""
 
 msgid "Save changes"
@@ -75,21 +90,6 @@ msgid "Execute"
 msgstr ""
 
 msgid "Run now"
-msgstr ""
-
-msgid "Click to select"
-msgstr ""
-
-msgid "Parameters"
-msgstr ""
-
-msgid "Select frequency"
-msgstr ""
-
-msgid "Cron expression"
-msgstr ""
-
-msgid "Continuous execution"
 msgstr ""
 
 msgid "Successfully deleted job"
