+++ src/components/JobFields/index.js
@@ -1,3 +1,5 @@
 import Type from './Type';
+import Name from './Name';
+import Cron from './Cron';
 
-export { Type };
+export { Type, Name, Cron };
