--- src/components/Schedule/index.js
@@ -1,1 +1,0 @@
-export { default } from './Schedule';
